ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
kafka-python==2.0.2
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
scikit-learn==1.4.0
redis==5.0.1
python-dotenv==1.0.0